Add available programs to BMW driving center scraper
Add all available programs for searching to the BMW driving center scraper. This makes it possible to search for any program on the website.

No breaking changes are introduced in this commit.

This commit adds the list of available programs to the BMW driving center scraper, making it easier to search for any program on the website. The list of available programs is stored in the `AVAILABLE_PROGRAMS` list, which can be easily modified to add or remove programs as needed.

None.